pyproject.toml
[project]
name = "boletin7"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.8",
    "jinja2>=3.1.6",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.47",
    "uvicorn[standard]>=0.40.0",
]
